cpp file for CommunicatorMPI
Move implementations of the communicator into a cpp file.
